+++ Source/Pe/Pe.Library.Database/IDatabaseTransaction.cs
@@ -7,7 +7,7 @@ namespace ContentTypeTextNet.Pe.Library.Database
     /// データベース実装におけるトランザクション処理。
     /// </summary>
     /// <remarks>
-    /// <para>これが実体化されてればトランザクション中でしょうね。</para>
+    /// <para><see cref="IDatabaseAccessor.BeginReadOnlyTransaction"/>でも説明のある通りデータベース実装のトランザクションではない可能性あり。</para>
     /// <para>コミット・ロールバックは一回だけ実行される使用を想定している。</para>
     /// </remarks>
     public interface IDatabaseTransaction: IDatabaseContext, IDisposable
@@ -26,4 +26,5 @@ namespace ContentTypeTextNet.Pe.Library.Database
 
         #endregion
     }
+
 }
